map some alarms to already defined battery alarms
This renames the OverTemperature / OverVoltage alarms to be the same as the
existing battery service. Maintanance is renamed to MaintanceNeeded, just for
clarity and on par with MaintanceActive. The overCurrent alarm is normally split
in HighChargeCurrent / HighDischargeCurrent but redflow maps these to the same
bit. Hence it is left as overCurrent, at least for now.